src/main/java/com/groupon/jenkins/buildtype/install_packages/buildconfiguration/configvalue/ListValue.java
/*
The MIT License (MIT)

Copyright (c) 2014, Groupon, Inc.

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
THE SOFTWARE.
 */
package com.groupon.jenkins.buildtype.install_packages.buildconfiguration.configvalue;

import java.util.ArrayList;
import java.util.Collection;
import java.util.Collections;
import java.util.List;

public class ListValue<T> extends ConfigValue<List<T>> {

    public ListValue(Object config) {
        super(config == null ? Collections.emptyList() : config);
    }

    @Override
    public void append(ConfigValue<?> otherConfig) {
        ArrayList<T> mergedList = new ArrayList<T>();
        mergedList.addAll(getValue());
        mergedList.addAll((Collection<? extends T>) otherConfig.getValue());
        setValue(mergedList);
    }

    @Override
    public <R> R getValue(Class<R> returnType) {
        return (R) getValue();
    }

    @Override
    public boolean isEmpty() {
        return getValue().isEmpty();
    }
}
