src/test/java/com/groupon/jenkins/buildtype/install_packages/buildconfiguration/configvalue/ListValueTest.java
/*
The MIT License (MIT)

Copyright (c) 2014, Groupon, Inc.

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
THE SOFTWARE.
 */
package com.groupon.jenkins.buildtype.install_packages.buildconfiguration.configvalue;

import java.util.Arrays;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;

public class ListValueTest {

    @Test
    public void should_append_to_existing_values() {
        ListValue<String> configValue = new ListValue<String>(Arrays.asList("one", "two"));
        ListValue<String> configValueToBeAppended = new ListValue<String>(Arrays.asList("three"));
        configValue.append(configValueToBeAppended);
        assertEquals(3, configValue.getValue().size());
        assertEquals("one", configValue.getValue().get(0));
        assertEquals("two", configValue.getValue().get(1));
        assertEquals("three", configValue.getValue().get(2));
    }

    @Test
    public void null_value_should_be_empty() {
        assertTrue(new ListValue<Object>(null).isEmpty());
    }
}
